Add manifest module
